# src/aspnetcore/src/Org.OpenAPITools/wwwroot/openapi-original.json
{
  "openapi" : "3.0.3",
  "info" : {
    "contact" : {
      "email" : "support@salestim.com",
      "name" : "SalesTim Support",
      "url" : "https://developers.salestim.com/api"
    },
    "description" : "The Microsoft Teams Governance API\n",
    "license" : {
      "name" : "(c) Copyright 2019-2020 SalesTim - all rights reserved.",
      "url" : "https://www.salestim.com/legal/eula"
    },
    "termsOfService" : "https://www.salestim.com/legal/tos",
    "title" : "SalesTim API",
    "version" : "1.0.0"
  },
  "servers" : [ {
    "description" : "SalesTim API v1.0 (Production environment, uses live data)",
    "url" : "https://api.salestim.io/v1.0"
  } ],
...
  "paths" : {
    "/store/templates" : {
      "get" : {
        "description" : "<a href=\"\" target=\"_blank\"> <img src=\"https://img.shields.io/badge/dynamic/json?url=https://api.salestim.io/v1.0/store/templates&label=Template%20Store&query=$.body.length&color=darkslateblue&suffix=%20templates%20published&logo=microsoft-teams&logoColor=white\" /> </a> <hr /> **View** all templates from the [Public Template Store](https://store.salestim.com) or **Publish** your own templates to the [Template Manifests Repository](https://github.com/SalesTim/template-manifests)\n",
        "externalDocs" : {
          "url" : "https://developers.salestim.com/api/reference/resources/store/#getStoreTemplates"
        },
        "operationId" : "GetStoreTemplates",
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ArrayOfStoreTemplates"
                }
              }
            },
            "description" : "Operation executed successfully."
          },
          "404" : {
            "description" : "The ressource cannot be found."
          },
          "429" : {
            "description" : "Too many requests. Your request has been throttled."
          },
          "401" : {
            "description" : "API key or app id are missing or invalid."
          },
          "500" : {
            "description" : "The server encountered an unexpected condition that prevented it from fulfilling the request."
          },
          "default" : {
            "description" : "The server encountered an unexpected condition that prevented it from fulfilling the request."
          }
        },
        "summary" : "Get all templates from the public template store",
        "tags" : [ "Store" ]
      }
    },
    "/store/templates/{id}" : {
      "get" : {
        "description" : "<a href=\"\" target=\"_blank\"> <img src=\"https://img.shields.io/badge/dynamic/json?url=https://api.salestim.io/v1.0/store/templates&label=Template%20Store&query=$.body.length&color=darkslateblue&suffix=%20templates%20published&logo=microsoft-teams&logoColor=white\" /> </a> <hr /> **View** all templates from the [Public Template Store](https://store.salestim.com) or **Publish** your own templates to the [Template Manifests Repository](https://github.com/SalesTim/template-manifests)\n",
        "externalDocs" : {
          "url" : "https://developers.salestim.com/api/reference/resources/store/#getStoreTemplate"
        },
        "operationId" : "GetStoreTemplate",
        "parameters" : [ {
          "description" : "The template ID",
          "explode" : false,
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "type" : "string"
          },
          "style" : "simple"
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/StoreTemplate"
                }
              }
            },
            "description" : "Operation executed successfully."
          },
          "404" : {
            "description" : "The ressource cannot be found."
          },
          "429" : {
            "description" : "Too many requests. Your request has been throttled."
          },
          "401" : {
            "description" : "API key or app id are missing or invalid."
          },
          "500" : {
            "description" : "The server encountered an unexpected condition that prevented it from fulfilling the request."
          },
          "default" : {
            "description" : "The server encountered an unexpected condition that prevented it from fulfilling the request."
          }
        },
        "summary" : "Get a store template",
        "tags" : [ "Store" ]
      }
    },
    "/store/templates/byCategory/{id}" : {
      "get" : {
        "description" : "<a href=\"\" target=\"_blank\"> <img src=\"https://img.shields.io/badge/dynamic/json?url=https://api.salestim.io/v1.0/store/templates&label=Template%20Store&query=$.body.length&color=darkslateblue&suffix=%20templates%20published&logo=microsoft-teams&logoColor=white\" /> </a> <hr /> **View** all templates from the [Public Template Store](https://store.salestim.com) or **Publish** your own templates to the [Template Manifests Repository](https://github.com/SalesTim/template-manifests)\n",
        "externalDocs" : {
          "url" : "https://developers.salestim.com/api/reference/resources/store/#getStoreTemplatesByCategory"
        },
        "operationId" : "GetStoreTemplatesByCategory",
        "parameters" : [ {
          "description" : "The category ID",
          "explode" : false,
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "type" : "string"
          },
          "style" : "simple"
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ArrayOfStoreTemplates"
                }
              }
            },
            "description" : "Operation executed successfully."
          },
          "404" : {
            "description" : "The ressource cannot be found."
          },
          "429" : {
            "description" : "Too many requests. Your request has been throttled."
          },
          "401" : {
            "description" : "API key or app id are missing or invalid."
          },
          "500" : {
            "description" : "The server encountered an unexpected condition that prevented it from fulfilling the request."
          },
          "default" : {
            "description" : "The server encountered an unexpected condition that prevented it from fulfilling the request."
          }
        },
        "summary" : "Get store template from a specific category",
        "tags" : [ "Store" ]
      }
    },
    "/store/categories" : {
      "get" : {
        "description" : "<a href=\"\" target=\"_blank\"> <img src=\"https://img.shields.io/badge/dynamic/json?url=https://api.salestim.io/v1.0/store/templates&label=Template%20Store&query=$.body.length&color=darkslateblue&suffix=%20templates%20published&logo=microsoft-teams&logoColor=white\" /> </a> <hr /> **View** all templates from the [Public Template Store](https://store.salestim.com) or **Publish** your own templates to the [Template Manifests Repository](https://github.com/SalesTim/template-manifests)\n",
        "externalDocs" : {
          "url" : "https://developers.salestim.com/api/reference/resources/store/#getStoreCategories"
        },
        "operationId" : "GetStoreCategories",
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ArrayOfStoreCategories"
                }
              }
            },
            "description" : "Operation executed successfully."
          },
          "404" : {
            "description" : "The ressource cannot be found."
          },
          "429" : {
            "description" : "Too many requests. Your request has been throttled."
          },
          "401" : {
            "description" : "API key or app id are missing or invalid."
          },
          "500" : {
            "description" : "The server encountered an unexpected condition that prevented it from fulfilling the request."
          },
          "default" : {
            "description" : "The server encountered an unexpected condition that prevented it from fulfilling the request."
          }
        },
        "summary" : "Get all store categories from the public template store",
        "tags" : [ "Store" ]
      }
    },
...
  },
  "components" : {
...
    "schemas" : {
      "ArrayOfStoreTemplates" : {
        "description" : "A list of templates from the public template store",
        "externalDocs" : {
          "url" : "https://developers.salestim.com/api/reference/resources/store"
        },
        "items" : {
          "$ref" : "#/components/schemas/StoreTemplate"
        },
        "type" : "array"
      },
      "StoreTemplate" : {
        "description" : "A template from the public template store",
        "example" : {
          "templateConfiguration" : {
            "id" : "id"
          }
        },
        "externalDocs" : {
          "url" : "https://developers.salestim.com/api/reference/resources/store"
        },
        "properties" : {
          "templateConfiguration" : {
            "$ref" : "#/components/schemas/StoreTemplate_templateConfiguration"
          }
        }
      },
      "ArrayOfStoreCategories" : {
        "description" : "A list of categories from the public template store",
        "externalDocs" : {
          "url" : "https://developers.salestim.com/api/reference/resources/store"
        },
        "items" : {
          "$ref" : "#/components/schemas/StoreCategory"
        },
        "type" : "array"
      },
      "StoreCategory" : {
        "description" : "A category of templates from the public template store",
        "example" : {
          "id" : "id",
          "label" : {
            "en-us" : "en-us",
            "fr-fr" : "fr-fr"
          },
          "slug" : "slug"
        },
        "externalDocs" : {
          "url" : "https://developers.salestim.com/api/reference/resources/store"
        },
        "properties" : {
          "id" : {
            "description" : "Category ID",
            "type" : "string"
          },
          "label" : {
            "$ref" : "#/components/schemas/StoreCategory_label"
          },
          "slug" : {
            "description" : "Category slug used for clean and seo-friendly urls",
            "type" : "string"
          }
        }
      },
...
      "StoreTemplate_templateConfiguration" : {
        "description" : "Main template properties",
        "example" : {
          "id" : "id"
        },
        "properties" : {
          "id" : {
            "description" : "Template ID",
            "type" : "string",
            "x-ms-summary" : "Template ID"
          }
        }
      },
      "StoreCategory_label" : {
        "description" : "Collection of localized category labels",
        "example" : {
          "en-us" : "en-us",
          "fr-fr" : "fr-fr"
        },
        "properties" : {
          "en-us" : {
            "description" : "Category label localized for en-us.",
            "type" : "string"
          },
          "fr-fr" : {
            "description" : "Category label localized for fr-fr.",
            "type" : "string"
          }
        }
      },
...
    },
    "securitySchemes" : {
      "bearerAuth" : {
        "bearerFormat" : "JWT",
        "description" : "A valid Microsoft 365 access_token issued by Azure AD",
        "scheme" : "bearer",
        "type" : "http"
      },
      "pluginId" : {
        "description" : "A valid plugin ID",
        "in" : "header",
        "name" : "X-APP-ID",
        "type" : "apiKey"
      },
      "pluginSecret" : {
        "description" : "A valid plugin secret",
        "in" : "header",
        "name" : "X-API-KEY",
        "type" : "apiKey"
      }
    }
  }
}